refactor: handler 함수 분리
handleMouseDown, handleMouseLeave, handleMouseMove
